requirements.txt
absl-py==0.11.0
arabic-reshaper==2.1.3
asgiref==3.4.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
chdir==1.0.0
click==7.1.2
comtypes==1.1.8
defusedxml==0.7.1
details==0.2.0
diff-match-patch==20200713
Django==3.2.5
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
excelReader==0.0.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gTTS==2.2.2
h5py==2.10.0
html5lib==1.1
idna==2.10
importlib-metadata==3.4.0
jdcal==1.4.1
json-excel-converter==1.3.0
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupPy==1.14
numpy==1.19.5
oauthlib==3.1.0
odfpy==1.4.1
Open-Tamil==0.98
opencv-python==4.5.1.48
openpyxl==3.0.6
opt-einsum==3.3.0
pandas==1.2.1
Pillow==8.3.1
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyPDF2==1.26.0
pypiwin32==223
python-bidi==0.4.2
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2021.1
pywin32==301
PyYAML==5.4.1
reportlab==3.5.68
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tf-slim==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
zipp==3.4.0
